move into vms.go
