+++ cmd/katapult/console/question.go
@@ -2,16 +2,17 @@ package console
 
 import (
 	"bufio"
-	"github.com/buger/goterm"
 	"io"
 	"strings"
+
+	"github.com/buger/goterm"
 )
 
 // Question is used to define a basic console question.
 func Question(question string, blankAcceptable bool, bufferedStdin *bufio.Reader, stdout io.Writer) string {
 	for {
 		// Print the question.
-		_, _ = stdout.Write([]byte(goterm.Color(question, goterm.CYAN)+" "))
+		_, _ = stdout.Write([]byte(goterm.Color(question, goterm.CYAN) + " "))
 
 		// Read stdin.
 		text, _ := bufferedStdin.ReadString('\n')
